Crate rustc_metadata

Unstable (rustc_private)

Modules

astencode 	 [Unstable]

common 	 [Unstable]

creader 	 [Unstable]
Validates all used crates and extern libraries and loads their metadata

csearch 	 [Unstable]

cstore 	 [Unstable]

decoder 	 [Unstable]

def_key 	 [Unstable]

diagnostics 	 [Unstable]

encoder 	 [Unstable]

index 	 [Unstable]

loader 	 [Unstable]
Finds crate binaries and loads their metadata

macro_import 	 [Unstable]
Used by rustc when loading a crate with exported macros.

middle 	 [Unstable]

tls_context 	 [Unstable]

tydecode 	 [Unstable]

tyencode 	 [Unstable]