Struct rustc_metadata::creader::CStore

pub struct CStore {
    metas: IndexVec<CrateNum, Option<Lrc<CrateMetadata>>>,
    injected_panic_runtime: Option<CrateNum>,
    allocator_kind: Option<AllocatorKind>,
    has_global_allocator: bool,
    pub(crate) stable_crate_ids: FxHashMap<StableCrateId, CrateNum>,
    unused_externs: Vec<Symbol>,
}

Fields
metas: IndexVec<CrateNum, Option<Lrc<CrateMetadata>>>injected_panic_runtime: Option<CrateNum>allocator_kind: Option<AllocatorKind>

This crate needs an allocator and either provides it itself, or finds it in a dependency. If the above is true, then this field denotes the kind of the found allocator.

has_global_allocator: bool

This crate has a #[global_allocator] item.

stable_crate_ids: FxHashMap<StableCrateId, CrateNum>

This map is used to verify we get no hash conflicts between StableCrateId values.

unused_externs: Vec<Symbol>

Unused externs of the crate

Implementations

impl CStore

pub fn struct_field_names_untracked<'a>(
&'a self,
def: DefId,
sess: &'a Session
) -> impl Iterator<Item = Spanned<Symbol>> + 'a

pub fn struct_field_visibilities_untracked(
&self,
def: DefId
) -> impl Iterator<Item = Visibility<DefId>> + '_

pub fn ctor_def_id_and_kind_untracked(
&self,
def: DefId
) -> Option<(DefId, CtorKind)>

pub fn visibility_untracked(&self, def: DefId) -> Visibility<DefId>

pub fn module_children_untracked(
&self,
def_id: DefId,
sess: &Session
) -> Vec<ModChild>

pub fn load_macro_untracked(&self, id: DefId, sess: &Session) -> LoadedMacro

pub fn fn_has_self_parameter_untracked(&self, def: DefId, sess: &Session) -> bool

pub fn crate_source_untracked(&self, cnum: CrateNum) -> Lrc<CrateSource>

pub fn get_span_untracked(&self, def_id: DefId, sess: &Session) -> Span

pub fn def_kind(&self, def: DefId) -> DefKind

pub fn crates_untracked(&self) -> impl Iterator<Item = CrateNum> + '_

pub fn item_generics_num_lifetimes(&self, def_id: DefId, sess: &Session) -> usize

pub fn module_expansion_untracked(&self, def_id: DefId, sess: &Session) -> ExpnId

pub fn num_def_ids_untracked(&self, cnum: CrateNum) -> usize

Only public-facing way to traverse all the definitions in a non-local crate. Critically useful for this third-party project: https://github.com/hacspec/hacspec. See https://github.com/rust-lang/rust/pull/85889 for context.

pub fn item_attrs_untracked<'a>(
&'a self,
def_id: DefId,
sess: &'a Session
) -> impl Iterator<Item = Attribute> + 'a

pub fn get_proc_macro_quoted_span_untracked(
&self,
cnum: CrateNum,
id: usize,
sess: &Session
) -> Span

pub fn traits_in_crate_untracked(
&self,
cnum: CrateNum
) -> impl Iterator<Item = DefId> + '_

Decodes all traits in the crate (for rustdoc).

pub fn trait_impls_in_crate_untracked(
&self,
cnum: CrateNum
) -> impl Iterator<Item = (DefId, DefId, Option<SimplifiedType>)> + '_

Decodes all trait impls in the crate (for rustdoc).

pub fn inherent_impls_in_crate_untracked(
&self,
cnum: CrateNum
) -> impl Iterator<Item = (DefId, DefId)> + '_

Decodes all inherent impls in the crate (for rustdoc).

pub fn incoherent_impls_in_crate_untracked(
&self,
cnum: CrateNum
) -> impl Iterator<Item = DefId> + '_

Decodes all incoherent inherent impls in the crate (for rustdoc).

pub fn associated_item_def_ids_untracked<'a>(
&'a self,
def_id: DefId,
sess: &'a Session
) -> impl Iterator<Item = DefId> + 'a

pub fn may_have_doc_links_untracked(&self, def_id: DefId) -> bool

impl CStore

pub fn from_tcx(tcx: TyCtxt<'_>) -> &CStore

fn alloc_new_crate_num(&mut self) -> CrateNum

pub fn has_crate_data(&self, cnum: CrateNum) -> bool

pub(crate) fn get_crate_data(&self, cnum: CrateNum) -> CrateMetadataRef<'_>

fn set_crate_data(&mut self, cnum: CrateNum, data: CrateMetadata)

pub(crate) fn iter_crate_data(
&self
) -> impl Iterator<Item = (CrateNum, &CrateMetadata)>

fn push_dependencies_in_postorder(&self, deps: &mut Vec<CrateNum>, cnum: CrateNum)

pub(crate) fn crate_dependencies_in_postorder(
&self,
cnum: CrateNum
) -> Vec<CrateNum>

fn crate_dependencies_in_reverse_postorder(
&self,
cnum: CrateNum
) -> Vec<CrateNum>

pub(crate) fn injected_panic_runtime(&self) -> Option<CrateNum>

pub(crate) fn allocator_kind(&self) -> Option<AllocatorKind>

pub(crate) fn has_global_allocator(&self) -> bool

pub fn report_unused_deps(&self, tcx: TyCtxt<'_>)