Struct rustc_metadata::creader::CrateLoader

pub struct CrateLoader<'a> {
    pub sess: &'a Session,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
sess: &'a Session

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a> CrateLoader<'a>

pub fn new(
sess: &'a Session,
cstore: &'a CStore,
local_crate_name: &str
) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn find_plugin_registrar(
&mut self,
span: Span,
name: &str
) -> Option<(PathBuf, CrateDisambiguator, DefIndex)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Look for a plugin registrar. Returns library path, crate SVH and DefIndex of the registrar function.