rustc_metadata::creader
Trait MetadataLoader

pub trait MetadataLoader {
    // Required methods
    fn get_rlib_metadata(
        &self,
        target: &Target,
        filename: &Path,
    ) -> Result<OwnedSlice, String>;
    fn get_dylib_metadata(
        &self,
        target: &Target,
        filename: &Path,
    ) -> Result<OwnedSlice, String>;
}

The backend’s way to give the crate store access to the metadata in a library. Note that it returns the raw metadata bytes stored in the library file, whether it is compressed, uncompressed, some weird mix, etc. rmeta files are backend independent and not handled here.

Required Methods

fn get_rlib_metadata( &self, target: &Target, filename: &Path, ) -> Result<OwnedSlice, String>

fn get_dylib_metadata( &self, target: &Target, filename: &Path, ) -> Result<OwnedSlice, String>