Struct rustc_metadata::cstore::CStore

pub struct CStore {
    pub dep_graph: DepGraph,
    pub dllimport_foreign_items: RefCell<FxHashSet<DefIndex>>,
    pub visible_parent_map: RefCell<DefIdMap<DefId>>,
    pub metadata_loader: Box<MetadataLoader>,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
dep_graph: DepGraph

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

dllimport_foreign_items: RefCell<FxHashSet<DefIndex>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

visible_parent_map: RefCell<DefIdMap<DefId>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

metadata_loader: Box<MetadataLoader>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl CStore

fn new(dep_graph: &DepGraph, metadata_loader: Box<MetadataLoader>) -> CStore

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn next_crate_num(&self) -> CrateNum

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_crate_data(&self, cnum: CrateNum) -> Rc<CrateMetadata>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_crate_hash(&self, cnum: CrateNum) -> Svh

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn set_crate_data(&self, cnum: CrateNum, data: Rc<CrateMetadata>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn iter_crate_data<I>(&self, i: I) where
I: FnMut(CrateNum, &Rc<CrateMetadata>),

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn crate_dependencies_in_rpo(&self, krate: CrateNum) -> Vec<CrateNum>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn push_dependencies_in_postorder(
&self,
ordering: &mut Vec<CrateNum>,
krate: CrateNum
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn do_get_used_crates(
&self,
prefer: LinkagePreference
) -> Vec<(CrateNum, LibSource)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn add_used_library(&self, lib: NativeLibrary)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_used_libraries(&self) -> &RefCell<Vec<NativeLibrary>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn add_used_link_args(&self, args: &str)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_used_link_args<'a>(&'a self) -> &'a RefCell<Vec<String>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn add_extern_mod_stmt_cnum(&self, emod_id: NodeId, cnum: CrateNum)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn add_statically_included_foreign_item(&self, id: DefIndex)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn do_is_statically_included_foreign_item(&self, def_id: DefId) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn do_extern_mod_stmt_cnum(&self, emod_id: NodeId) -> Option<CrateNum>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn read_dep_node(&self, def_id: DefId)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?