Struct rustc_metadata::cstore::CrateMetadata

pub struct CrateMetadata {
    pub name: Symbol,
    pub extern_crate: Cell<Option<ExternCrate>>,
    pub blob: MetadataBlob,
    pub cnum_map: RefCell<CrateNumMap>,
    pub cnum: CrateNum,
    pub codemap_import_info: RefCell<Vec<ImportedFileMap>>,
    pub attribute_cache: RefCell<[Vec<Option<Rc<[Attribute]>>>; 2]>,
    pub root: CrateRoot,
    pub def_path_table: Rc<DefPathTable>,
    pub exported_symbols: Tracked<FxHashSet<DefIndex>>,
    pub trait_impls: Tracked<FxHashMap<(u32, DefIndex), LazySeq<DefIndex>>>,
    pub dep_kind: Cell<DepKind>,
    pub source: CrateSource,
    pub proc_macros: Option<Vec<(Name, Rc<SyntaxExtension>)>>,
    pub dllimport_foreign_items: Tracked<FxHashSet<DefIndex>>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
name: Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

extern_crate: Cell<Option<ExternCrate>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Information about the extern crate that caused this crate to be loaded. If this is None, then the crate was injected (e.g., by the allocator)

blob: MetadataBlob

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

cnum_map: RefCell<CrateNumMap>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

cnum: CrateNum

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

codemap_import_info: RefCell<Vec<ImportedFileMap>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

attribute_cache: RefCell<[Vec<Option<Rc<[Attribute]>>>; 2]>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

root: CrateRoot

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

def_path_table: Rc<DefPathTable>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

For each public item in this crate, we encode a key. When the crate is loaded, we read all the keys and put them in this hashmap, which gives the reverse mapping. This allows us to quickly retrace a DefPath, which is needed for incremental compilation support.

exported_symbols: Tracked<FxHashSet<DefIndex>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

trait_impls: Tracked<FxHashMap<(u32, DefIndex), LazySeq<DefIndex>>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

dep_kind: Cell<DepKind>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

source: CrateSource

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

proc_macros: Option<Vec<(Name, Rc<SyntaxExtension>)>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

dllimport_foreign_items: Tracked<FxHashSet<DefIndex>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<'a, 'tcx> CrateMetadata

fn item_name(&self, item_index: DefIndex) -> Name

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_def(&self, index: DefIndex) -> Option<Def>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_span(&self, index: DefIndex, sess: &Session) -> Span

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_trait_def(&self, item_id: DefIndex) -> TraitDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_adt_def(
&self,
item_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>
) -> &'tcx AdtDef

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_predicates(
&self,
item_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>
) -> GenericPredicates<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_super_predicates(
&self,
item_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>
) -> GenericPredicates<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_generics(&self, item_id: DefIndex) -> Generics

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_type(&self, id: DefIndex, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> Ty<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_stability(&self, id: DefIndex) -> Option<Stability>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_deprecation(&self, id: DefIndex) -> Option<Deprecation>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_visibility(&self, id: DefIndex) -> Visibility

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_parent_impl(&self, id: DefIndex) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_impl_polarity(&self, id: DefIndex) -> ImplPolarity

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_impl_defaultness(&self, id: DefIndex) -> Defaultness

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_coerce_unsized_info(&self, id: DefIndex) -> Option<CoerceUnsizedInfo>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_impl_trait(
&self,
id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>
) -> Option<TraitRef<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_lang_items(&self, dep_graph: &DepGraph) -> Vec<(DefIndex, usize)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Iterates over the language items in the given crate.

fn each_child_of_item<F>(&self, id: DefIndex, callback: F, sess: &Session) where
F: FnMut(Export),

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Iterates over each child of the given item.

fn item_body(&self, tcx: TyCtxt<'a, 'tcx, 'tcx>, id: DefIndex) -> &'tcx Body

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn item_body_tables(
&self,
id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>
) -> &'tcx TypeckTables<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn item_body_nested_bodies(&self, id: DefIndex) -> BTreeMap<BodyId, Body>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn const_is_rvalue_promotable_to_static(&self, id: DefIndex) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_item_mir_available(&self, id: DefIndex) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn maybe_get_optimized_mir(
&self,
tcx: TyCtxt<'a, 'tcx, 'tcx>,
id: DefIndex
) -> Option<Mir<'tcx>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn mir_const_qualif(&self, id: DefIndex) -> u8

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_associated_item(&self, id: DefIndex) -> AssociatedItem

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_item_variances(&self, id: DefIndex) -> Vec<Variance>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_ctor_kind(&self, node_id: DefIndex) -> CtorKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_struct_ctor_def_id(&self, node_id: DefIndex) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_item_attrs(
&self,
node_id: DefIndex,
dep_graph: &DepGraph
) -> Rc<[Attribute]>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_struct_field_names(&self, id: DefIndex) -> Vec<Name>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_inherent_implementations_for_type(&self, id: DefIndex) -> Vec<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_implementations_for_trait(
&self,
filter: Option<DefId>,
dep_graph: &DepGraph,
result: &mut Vec<DefId>
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_trait_of_item(&self, id: DefIndex) -> Option<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_native_libraries(&self, dep_graph: &DepGraph) -> Vec<NativeLibrary>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_dylib_dependency_formats(
&self,
dep_graph: &DepGraph
) -> Vec<(CrateNum, LinkagePreference)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_missing_lang_items(&self, dep_graph: &DepGraph) -> Vec<LangItem>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_fn_arg_names(&self, id: DefIndex) -> Vec<Name>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_exported_symbols(&self, dep_graph: &DepGraph) -> Vec<DefId>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn get_macro(&self, id: DefIndex) -> (Name, MacroDef)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_const_fn(&self, id: DefIndex) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_foreign_item(&self, id: DefIndex) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_dllimport_foreign_item(&self, id: DefIndex, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_default_impl(&self, impl_id: DefIndex) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn closure_kind(&self, closure_id: DefIndex) -> ClosureKind

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn fn_sig(&self, id: DefIndex, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> PolyFnSig<'tcx>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn def_key(&self, index: DefIndex) -> DefKey

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn def_path(&self, id: DefIndex) -> DefPath

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn def_path_hash(&self, index: DefIndex) -> DefPathHash

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn imported_filemaps(
&'a self,
local_codemap: &CodeMap
) -> Ref<'a, Vec<ImportedFileMap>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Imports the codemap from an external crate into the codemap of the crate currently being compiled (the "local crate").

The import algorithm works analogous to how AST items are inlined from an external crate's metadata: For every FileMap in the external codemap an 'inline' copy is created in the local codemap. The correspondence relation between external and local FileMaps is recorded in the ImportedFileMap objects returned from this function. When an item from an external crate is later inlined into this crate, this correspondence information is used to translate the span information of the inlined item so that it refers the correct positions in the local codemap (see <decoder::DecodeContext as SpecializedDecoder<Span>>).

The import algorithm in the function below will reuse FileMaps already existing in the local codemap. For example, even if the FileMap of some source file of libstd gets imported many times, there will only ever be one FileMap object for the corresponding file in the local codemap.

Note that imported FileMaps do not actually contain the source code of the file they represent, just information about length, line breaks, and multibyte characters. This information is enough to generate valid debuginfo for items inlined from other crates.

fn metadata_dep_node(&self, kind: GlobalMetaDataKind) -> DepNode

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl CrateMetadata

fn name(&self) -> Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn hash(&self) -> Svh

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn disambiguator(&self) -> Symbol

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn needs_allocator(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn has_global_allocator(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn has_default_lib_allocator(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_panic_runtime(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn needs_panic_runtime(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_compiler_builtins(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_sanitizer_runtime(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_profiler_runtime(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn is_no_builtins(&self, dep_graph: &DepGraph) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn panic_strategy(&self, dep_graph: &DepGraph) -> PanicStrategy

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?