Struct rustc_metadata::cstore::ImportedFileMap

pub struct ImportedFileMap {
    pub original_start_pos: BytePos,
    pub original_end_pos: BytePos,
    pub translated_filemap: Rc<FileMap>,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Holds information about a syntax_pos::FileMap imported from another crate. See imported_filemaps() for more information.

Fields
original_start_pos: BytePos

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This FileMap's byte-offset within the codemap of its original crate

original_end_pos: BytePos

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The end of this FileMap within the codemap of its original crate

translated_filemap: Rc<FileMap>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

The imported FileMap's representation within the local codemap