Enum rustc_metadata::middle::const_val::ConstInt

pub enum ConstInt {
    I8(i8),
    I16(i16),
    I32(i32),
    I64(i64),
    I128(i128),
    Isize(ConstIsize),
    U8(u8),
    U16(u16),
    U32(u32),
    U64(u64),
    U128(u128),
    Usize(ConstUsize),
}

🔬 This is a nightly-only experimental API. (rustc_private)

Variants
I8(i8)

🔬 This is a nightly-only experimental API. (rustc_private)

I16(i16)

🔬 This is a nightly-only experimental API. (rustc_private)

I32(i32)

🔬 This is a nightly-only experimental API. (rustc_private)

I64(i64)

🔬 This is a nightly-only experimental API. (rustc_private)

I128(i128)

🔬 This is a nightly-only experimental API. (rustc_private)

Isize(ConstIsize)

🔬 This is a nightly-only experimental API. (rustc_private)

U8(u8)

🔬 This is a nightly-only experimental API. (rustc_private)

U16(u16)

🔬 This is a nightly-only experimental API. (rustc_private)

U32(u32)

🔬 This is a nightly-only experimental API. (rustc_private)

U64(u64)

🔬 This is a nightly-only experimental API. (rustc_private)

U128(u128)

🔬 This is a nightly-only experimental API. (rustc_private)

Usize(ConstUsize)

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl ConstInt

fn new_unsigned(val: u128, ty: UintTy, usize_ty: UintTy) -> Option<ConstInt>

🔬 This is a nightly-only experimental API. (rustc_private)

Creates a new unsigned ConstInt with matching type while also checking that overflow does not happen.

fn new_signed(val: i128, ty: IntTy, isize_ty: IntTy) -> Option<ConstInt>

🔬 This is a nightly-only experimental API. (rustc_private)

Creates a new signed ConstInt with matching type while also checking that overflow does not happen.

fn new_unsigned_truncating(val: u128, ty: UintTy, usize_ty: UintTy) -> ConstInt

🔬 This is a nightly-only experimental API. (rustc_private)

Creates a new unsigned ConstInt with matching type.

fn new_signed_truncating(val: i128, ty: IntTy, isize_ty: IntTy) -> ConstInt

🔬 This is a nightly-only experimental API. (rustc_private)

Creates a new signed ConstInt with matching type.

fn description(&self) -> &'static str

🔬 This is a nightly-only experimental API. (rustc_private)

Description of the type, not the value

fn to_u128_unchecked(self) -> u128

🔬 This is a nightly-only experimental API. (rustc_private)

Erases the type and returns a u128. This is not the same as -5i8 as u128 but as -5i8 as i128 as u128

fn to_u32(&self) -> Option<u32>

🔬 This is a nightly-only experimental API. (rustc_private)

Converts the value to a u32 if it's in the range 0...std::u32::MAX

fn to_u64(&self) -> Option<u64>

🔬 This is a nightly-only experimental API. (rustc_private)

Converts the value to a u64 if it's in the range 0...std::u64::MAX

fn to_u128(&self) -> Option<u128>

🔬 This is a nightly-only experimental API. (rustc_private)

Converts the value to a u128 if it's in the range 0...std::u128::MAX

fn to_f32(self) -> f32

🔬 This is a nightly-only experimental API. (rustc_private)

fn to_f64(self) -> f64

🔬 This is a nightly-only experimental API. (rustc_private)

fn is_negative(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

fn try_cmp(self, rhs: ConstInt) -> Result<Ordering, ConstMathErr>

🔬 This is a nightly-only experimental API. (rustc_private)

Compares the values if they are of the same type

fn wrap_incr(self) -> ConstInt

🔬 This is a nightly-only experimental API. (rustc_private)

Adds 1 to the value and wraps around if the maximum for the type is reached

fn int_type(self) -> IntType

🔬 This is a nightly-only experimental API. (rustc_private)