Module rustc_metadata::middle::cstore::tls

Unstable (rustc_private)

Metadata encoding and decoding can make use of thread-local encoding and decoding contexts. These allow implementers of serialize::Encodable and Decodable to access information and datastructures that would otherwise not be available to them. For example, we can automatically translate def-id and span information during decoding because the decoding context knows which crate the data is decoded from. Or it allows to make ty::Ty decodable because the context has access to the TyCtxt that is needed for creating ty::Ty instances.

Note, however, that this only works for RBML-based encoding and decoding at the moment.

Traits

DecodingContext 	 [Unstable]

EncodingContext 	 [Unstable]

Functions

enter_decoding_context 	 [Unstable]
Execute f after pushing the given DecodingContext onto the TLS stack.

enter_encoding_context 	 [Unstable]
Execute f after pushing the given EncodingContext onto the TLS stack.

unsafe_with_decoding_context 	 [Unstable]
Execute f with access to the thread-local decoding context and rbml decoder.

unsafe_with_encoding_context 	 [Unstable]
Execute f with access to the thread-local encoding context and rbml encoder.

with_decoding_context 	 [Unstable]
Execute f with access to the thread-local decoding context and rbml decoder. This function will panic if the decoder passed in and the context decoder are not the same.

with_encoding_context 	 [Unstable]
Execute f with access to the thread-local encoding context and rbml encoder. This function will panic if the encoder passed in and the context encoder are not the same.