Struct rustc_metadata::middle::dataflow::DataFlowContext

pub struct DataFlowContext<'a, 'tcx, O> where 'tcx: 'a { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

Methods

impl<'a, 'tcx, O> DataFlowContext<'a, 'tcx, O> where O: DataFlowOperator

fn new(tcx: TyCtxt<'a, 'tcx, 'tcx>,
analysis_name: &'static str,
body: Option<&Body>,
cfg: &CFG,
oper: O,
id_range: IdRange,
bits_per_id: usize)
-> DataFlowContext<'a, 'tcx, O>

🔬 This is a nightly-only experimental API. (rustc_private)

fn add_gen(&mut self, id: NodeId, bit: usize)

🔬 This is a nightly-only experimental API. (rustc_private)

Indicates that id generates bit

fn add_kill(&mut self, kind: KillFrom, id: NodeId, bit: usize)

🔬 This is a nightly-only experimental API. (rustc_private)

Indicates that id kills bit

fn each_bit_on_entry<F>(&self, id: NodeId, f: F) -> bool where F: FnMut(usize) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Iterates through each bit that is set on entry to id. Only useful after propagate() has been called.

fn each_bit_for_node<F>(&self, e: EntryOrExit, cfgidx: NodeIndex, f: F) -> bool where F: FnMut(usize) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Iterates through each bit that is set on entry/exit to cfgidx. Only useful after propagate() has been called.

fn each_gen_bit<F>(&self, id: NodeId, f: F) -> bool where F: FnMut(usize) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

Iterates through each bit in the gen set for id.

fn add_kills_from_flow_exits(&mut self, cfg: &CFG)

🔬 This is a nightly-only experimental API. (rustc_private)

Whenever you have a break or continue statement, flow exits through any number of enclosing scopes on its way to the new destination. This function infers the kill bits of those control operators based on the kill bits associated with those scopes.

This is usually called (if it is called at all), after all add_gen and add_kill calls, but before propagate.

impl<'a, 'tcx, O> DataFlowContext<'a, 'tcx, O> where O: 'static + DataFlowOperator + Clone

fn propagate(&mut self, cfg: &CFG, body: &Body)

🔬 This is a nightly-only experimental API. (rustc_private)

Performs the data flow analysis.