Function rustc_metadata::middle::traits::type_known_to_meet_builtin_bound

pub fn type_known_to_meet_builtin_bound(infcx: &InferCtxt<'a, 'tcx>, ty: &'tcx TyS<'tcx>, bound: BuiltinBound, span: Span) -> bool

Unstable (rustc_private)

Determines whether the type ty is known to meet bound and returns true if so. Returns false if ty either does not meet bound or is not known to meet bound (note that this is conservative towards no impl, which is the opposite of the evaluate methods).