Enum rustc_metadata::middle::ty::TypeVariants

pub enum TypeVariants<'tcx> {
    TyBool,
    TyChar,
    TyInt(IntTy),
    TyUint(UintTy),
    TyFloat(FloatTy),
    TyEnum(&'tcx AdtDefData<'tcx, 'static>, &'tcx Substs<'tcx>),
    TyStruct(&'tcx AdtDefData<'tcx, 'static>, &'tcx Substs<'tcx>),
    TyBox(&'tcx TyS<'tcx>),
    TyStr,
    TyArray(&'tcx TyS<'tcx>, usize),
    TySlice(&'tcx TyS<'tcx>),
    TyRawPtr(TypeAndMut<'tcx>),
    TyRef(&'tcx Region, TypeAndMut<'tcx>),
    TyFnDef(DefId, &'tcx Substs<'tcx>, &'tcx BareFnTy<'tcx>),
    TyFnPtr(&'tcx BareFnTy<'tcx>),
    TyTrait(Box<TraitTy<'tcx>>),
    TyClosure(DefId, Box<ClosureSubsts<'tcx>>),
    TyTuple(Vec<&'tcx TyS<'tcx>>),
    TyProjection(ProjectionTy<'tcx>),
    TyParam(ParamTy),
    TyInfer(InferTy),
    TyError,
}

Unstable (rustc_private)

Variants

TyBool
Unstable (rustc_private)

The primitive boolean type. Written as bool.

TyChar
Unstable (rustc_private)

The primitive character type; holds a Unicode scalar value (a non-surrogate code point). Written as char.

TyInt
Unstable (rustc_private)

A primitive signed integer type. For example, i32.

TyUint
Unstable (rustc_private)

A primitive unsigned integer type. For example, u32.

TyFloat
Unstable (rustc_private)

A primitive floating-point type. For example, f64.

TyEnum
Unstable (rustc_private)

An enumerated type, defined with enum.

Substs here, possibly against intuition, may contain TyParams. That is, even after substitution it is possible that there are type variables. This happens when the TyEnum corresponds to an enum definition and not a concrete use of it. To get the correct TyEnum from the tcx, use the NodeId from the ast::Ty and look it up in the ast_ty_to_ty_cache. This is probably true for TyStruct as well.

TyStruct
Unstable (rustc_private)

A structure type, defined with struct.

See warning about substitutions for enumerated types.

TyBox
Unstable (rustc_private)

Box<T>; this is nominally a struct in the documentation, but is special-cased internally. For example, it is possible to implicitly move the contents of a box out of that box, and methods of any type can have type Box<Self>.

TyStr
Unstable (rustc_private)

The pointee of a string slice. Written as str.

TyArray
Unstable (rustc_private)

An array with the given length. Written as [T; n].

TySlice
Unstable (rustc_private)

The pointee of an array slice. Written as [T].

TyRawPtr
Unstable (rustc_private)

A raw pointer. Written as *mut T or *const T

TyRef
Unstable (rustc_private)

A reference; a pointer with an associated lifetime. Written as &a mut T or &'a T.

TyFnDef
Unstable (rustc_private)

The anonymous type of a function declaration/definition. Each function has a unique type.

TyFnPtr
Unstable (rustc_private)

A pointer to a function. Written as fn() -> i32. FIXME: This is currently also used to represent the callee of a method; see ty::MethodCallee etc.

TyTrait
Unstable (rustc_private)

A trait, defined with trait.

TyClosure
Unstable (rustc_private)

The anonymous type of a closure. Used to represent the type of |a| a.

TyTuple
Unstable (rustc_private)

A tuple type. For example, (i32, bool).

TyProjection
Unstable (rustc_private)

The projection of an associated type. For example, <T as Trait<..>>::N.

TyParam
Unstable (rustc_private)

A type parameter; for example, T in `fn f(x: T) {}

TyInfer
Unstable (rustc_private)

A type variable used during type-checking.

TyError
Unstable (rustc_private)

A placeholder for a type which could not be computed; this is propagated to avoid useless error messages.