Struct rustc_metadata::middle::ty::fold::RegionFolder

pub struct RegionFolder<'a, 'tcx> where 'tcx: 'a {
    // some fields omitted
}

Unstable (rustc_private)

Folds over the substructure of a type, visiting its component types and all regions that occur free within it.

That is, Ty can contain function or method types that bind regions at the call site (ReLateBound), and occurrences of regions (aka "lifetimes") that are bound within a type are not visited by this folder; only regions that occur free will be visited by fld_r.

Methods

impl<'a, 'tcx> RegionFolder<'a, 'tcx>

fn new<F>(tcx: &'a ctxt<'tcx>, skipped_regions: &'a mut bool, fld_r: &'a mut F) -> RegionFolder<'a, 'tcx> where F: FnMut(Region, u32) -> Region

Unstable (rustc_private)