Struct rustc_metadata::middle::ty::ExistentialBounds

pub struct ExistentialBounds<'tcx> {
    pub region_bound: Region,
    pub builtin_bounds: BuiltinBounds,
    pub projection_bounds: Vec<Binder<ProjectionPredicate<'tcx>>>,
}

Unstable (rustc_private)

Bounds suitable for an existentially quantified type parameter such as those that appear in object types or closure types.

Fields

region_bound
Unstable (rustc_private)

builtin_bounds
Unstable (rustc_private)

projection_bounds
Unstable (rustc_private)

Methods

impl<'tcx> ExistentialBounds<'tcx>

fn new(region_bound: Region, builtin_bounds: BuiltinBounds, projection_bounds: Vec<Binder<ProjectionPredicate<'tcx>>>) -> ExistentialBounds<'tcx>

Unstable (rustc_private)