Struct rustc_metadata::middle::ty::TypeContents

pub struct TypeContents {
    pub bits: u64,
}

Unstable (rustc_private)

Type contents is how the type checker reasons about kinds. They track what kinds of things are found within a type. You can think of them as kind of an "anti-kind". They track the kinds of values and thinks that are contained in types. Having a larger contents for a type tends to rule that type out from various kinds. For example, a type that contains a reference is not sendable.

The reason we compute type contents and not kinds is that it is easier for me (nmatsakis) to think about what is contained within a type than to think about what is not contained within a type.

Fields

bits
Unstable (rustc_private)

Methods

impl TypeContents

fn when(&self, cond: bool) -> TypeContents

Unstable (rustc_private)

fn intersects(&self, tc: TypeContents) -> bool

Unstable (rustc_private)

fn owns_owned(&self) -> bool

Unstable (rustc_private)

fn interior_param(&self) -> bool

Unstable (rustc_private)

fn interior_unsafe(&self) -> bool

Unstable (rustc_private)

fn needs_drop(&self, &ctxt) -> bool

Unstable (rustc_private)

fn owned_pointer(&self) -> TypeContents

Unstable (rustc_private)

Includes only those bits that still apply when indirected through a Box pointer

fn union<T, F>(v: &[T], f: F) -> TypeContents where F: FnMut(&T) -> TypeContents

Unstable (rustc_private)

fn has_dtor(&self) -> bool

Unstable (rustc_private)