Enum rustc_metadata::middle::ty::util::Representability

pub enum Representability {
    Representable,
    ContainsRecursive,
    SelfRecursive,
}

Unstable (rustc_private)

Describes whether a type is representable. For types that are not representable, 'SelfRecursive' and 'ContainsRecursive' are used to distinguish between types that are recursive with themselves and types that contain a different recursive type. These cases can therefore be treated differently when reporting errors.

The ordering of the cases is significant. They are sorted so that cmp::max will keep the "more erroneous" of two values.

Variants

Representable
Unstable (rustc_private)

ContainsRecursive
Unstable (rustc_private)

SelfRecursive
Unstable (rustc_private)