rustc_metadata::rmeta::decoder
Type Alias CrateNumMap

pub(crate) type CrateNumMap = IndexVec<CrateNum, CrateNum>;

A map from external crate numbers (as decoded from some crate file) to local crate numbers (as generated during this session). Each external crate may refer to types in other external crates, and each has their own crate numbers.

Aliased Type

struct CrateNumMap {
    pub raw: Vec<CrateNum>,
    /* private fields */
}

Fields
raw: Vec<CrateNum>
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 24 bytes