Struct rustc_metadata::rmeta::CrateRoot

pub(crate) struct CrateRoot {

    header: CrateHeader,
    extra_filename: String,
    stable_crate_id: StableCrateId,
    required_panic_strategy: Option<PanicStrategy>,
    panic_in_drop_strategy: PanicStrategy,
    edition: Edition,
    has_global_allocator: bool,
    has_alloc_error_handler: bool,
    has_panic_handler: bool,
    has_default_lib_allocator: bool,
    crate_deps: LazyArray<CrateDep>,
    dylib_dependency_formats: LazyArray<Option<LinkagePreference>>,
    lib_features: LazyArray<(Symbol, Option<Symbol>)>,
    stability_implications: LazyArray<(Symbol, Symbol)>,
    lang_items: LazyArray<(DefIndex, LangItem)>,
    lang_items_missing: LazyArray<LangItem>,
    stripped_cfg_items: LazyArray<StrippedCfgItem<DefIndex>>,
    diagnostic_items: LazyArray<(Symbol, DefIndex)>,
    native_libraries: LazyArray<NativeLib>,
    foreign_modules: LazyArray<ForeignModule>,
    traits: LazyArray<DefIndex>,
    impls: LazyArray<TraitImpls>,
    incoherent_impls: LazyArray<IncoherentImpls>,
    interpret_alloc_index: LazyArray<u64>,
    proc_macro_data: Option<ProcMacroData>,
    tables: LazyTables,
    debugger_visualizers: LazyArray<DebuggerVisualizerFile>,
    exported_symbols: LazyArray<(ExportedSymbol<'static>, SymbolExportInfo)>,
    syntax_contexts: LazyTable<u32, Option<LazyValue<SyntaxContextData>>>,
    expn_data: LazyTable<ExpnIndex, Option<LazyValue<ExpnData>>>,
    expn_hashes: LazyTable<ExpnIndex, Option<LazyValue<ExpnHash>>>,
    def_path_hash_map: LazyValue<DefPathHashMapRef<'static>>,
    source_map: LazyTable<u32, Option<LazyValue<SourceFile>>>,
    compiler_builtins: bool,
    needs_allocator: bool,
    needs_panic_runtime: bool,
    no_builtins: bool,
    panic_runtime: bool,
    profiler_runtime: bool,
    symbol_mangling_version: SymbolManglingVersion,

}

Serialized .rmeta data for a crate.

When compiling a proc-macro crate, we encode many of the LazyArray<T> fields as Lazy::empty(). This serves two purposes:

We avoid performing unnecessary work. Proc-macro crates can only export proc-macros functions, which are compiled into a shared library. As a result, a large amount of the information we normally store (e.g. optimized MIR) is unneeded by downstream crates.

We avoid serializing invalid CrateNums. When we deserialize a proc-macro crate, we don’t load any of its dependencies (since we just need to invoke a native function from the shared library). This means that any foreign CrateNums that we serialize cannot be deserialized, since we will not know how to map them into the current compilation session. If we were to serialize a proc-macro crate like a normal crate, much of what we serialized would be unusable in addition to being unused.

Fields
header: CrateHeader

A header used to detect if this is the right crate to load.

extra_filename: Stringstable_crate_id: StableCrateIdrequired_panic_strategy: Option<PanicStrategy>panic_in_drop_strategy: PanicStrategyedition: Editionhas_global_allocator: boolhas_alloc_error_handler: boolhas_panic_handler: boolhas_default_lib_allocator: boolcrate_deps: LazyArray<CrateDep>dylib_dependency_formats: LazyArray<Option<LinkagePreference>>lib_features: LazyArray<(Symbol, Option<Symbol>)>stability_implications: LazyArray<(Symbol, Symbol)>lang_items: LazyArray<(DefIndex, LangItem)>lang_items_missing: LazyArray<LangItem>stripped_cfg_items: LazyArray<StrippedCfgItem<DefIndex>>diagnostic_items: LazyArray<(Symbol, DefIndex)>native_libraries: LazyArray<NativeLib>foreign_modules: LazyArray<ForeignModule>traits: LazyArray<DefIndex>impls: LazyArray<TraitImpls>incoherent_impls: LazyArray<IncoherentImpls>interpret_alloc_index: LazyArray<u64>proc_macro_data: Option<ProcMacroData>tables: LazyTablesdebugger_visualizers: LazyArray<DebuggerVisualizerFile>exported_symbols: LazyArray<(ExportedSymbol<'static>, SymbolExportInfo)>syntax_contexts: LazyTable<u32, Option<LazyValue<SyntaxContextData>>>expn_data: LazyTable<ExpnIndex, Option<LazyValue<ExpnData>>>expn_hashes: LazyTable<ExpnIndex, Option<LazyValue<ExpnHash>>>def_path_hash_map: LazyValue<DefPathHashMapRef<'static>>source_map: LazyTable<u32, Option<LazyValue<SourceFile>>>compiler_builtins: boolneeds_allocator: boolneeds_panic_runtime: boolno_builtins: boolpanic_runtime: boolprofiler_runtime: boolsymbol_mangling_version: SymbolManglingVersion
Implementations

impl CrateRoot

pub(crate) fn is_proc_macro_crate(&self) -> bool

pub(crate) fn name(&self) -> Symbol

pub(crate) fn hash(&self) -> Svh

pub(crate) fn stable_crate_id(&self) -> StableCrateId

pub(crate) fn decode_crate_deps<'a>( &self, metadata: &'a MetadataBlob ) -> impl ExactSizeIterator<Item = CrateDep> + Captures<'a>