Struct rustc_metadata::rmeta::LazyTable

struct LazyTable<I, T> {
    position: NonZeroUsize,
    encoded_size: usize,
    _marker: PhantomData<fn(_: I) -> T>,
}

A list of lazily-decoded values, with the added capability of random access.

Random-access table (i.e. offering constant-time get/set), similar to LazyArray<T>, but without requiring encoding or decoding all the values eagerly and in-order.

Fields
position: NonZeroUsizeencoded_size: usize_marker: PhantomData<fn(_: I) -> T>
Implementations

impl<I: Idx, T: ParameterizedOverTcx> LazyTable<I, T>where
Option<T>: FixedSizeEncoding,

pub(super) fn get<'a, 'tcx, M: Metadata<'a, 'tcx>, const N: usize>(
&self,
metadata: M,
i: I
) -> Option<T::Value<'tcx>>where
Option<T::Value<'tcx>>: FixedSizeEncoding<ByteArray = [u8; N]>,

Given the metadata, extract out the value at a particular index (if any).

pub(super) fn size<const N: usize>(&self) -> usizewhere
for<'tcx> Option<T::Value<'tcx>>: FixedSizeEncoding<ByteArray = [u8; N]>,

Size of the table in entries, including possible gaps.

impl<I, T> LazyTable<I, T>

fn from_position_and_encoded_size(
position: NonZeroUsize,
encoded_size: usize
) -> LazyTable<I, T>