Type Alias rustc_metadata::rmeta::ExpnDataTable

type ExpnDataTable = LazyTable<ExpnIndex, Option<LazyValue<ExpnData>>>;

Aliased Type

struct ExpnDataTable {
    position: NonZero<usize>,
    width: usize,
    len: usize,
    _marker: PhantomData<fn(_: ExpnIndex) -> Option<LazyValue<ExpnData>>>,
}

Fields
position: NonZero<usize>width: usize

The encoded size of the elements of a table is selected at runtime to drop trailing zeroes. This is the number of bytes used for each table element.

len: usize

How many elements are in the table.

_marker: PhantomData<fn(_: ExpnIndex) -> Option<LazyValue<ExpnData>>>
Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 24 bytes