Type Alias rustc_metadata::rmeta::SyntaxContextTable

type SyntaxContextTable = LazyTable<u32, Option<LazyValue<SyntaxContextData>>>;

Aliased Type

struct SyntaxContextTable {
    position: NonZeroUsize,
    width: usize,
    len: usize,
    _marker: PhantomData<fn(_: u32) -> Option<LazyValue<SyntaxContextData>>>,
}

Fields
position: NonZeroUsizewidth: usize

The encoded size of the elements of a table is selected at runtime to drop trailing zeroes. This is the number of bytes used for each table element.

len: usize

How many elements are in the table.

_marker: PhantomData<fn(_: u32) -> Option<LazyValue<SyntaxContextData>>>
Implementations

impl<I: Idx, const N: usize, T: FixedSizeEncoding<ByteArray = [u8; N]> + ParameterizedOverTcx> LazyTable<I, T>where for<'tcx> T::Value<'tcx>: FixedSizeEncoding<ByteArray = [u8; N]>,

pub(super) fn get<'a, 'tcx, M: Metadata<'a, 'tcx>>( &self, metadata: M, i: I ) -> T::Value<'tcx>

Given the metadata, extract out the value at a particular index (if any).

pub(super) fn size(&self) -> usize

Size of the table in entries, including possible gaps.

impl<I, T> LazyTable<I, T>

fn from_position_and_encoded_size( position: NonZeroUsize, width: usize, len: usize ) -> LazyTable<I, T>