Crate rustc_metadata

Re-exports

pub use fs::emit_wrapper_file;

pub use fs::METADATA_FILENAME;

Modules

creader

Validates all used crates and extern libraries and loads their metadata

dependency_format 🔒

Resolution of mixing rlibs and dylibs

errors

foreign_modules 🔒

fs

locator

Finds crate binaries and loads their metadata

native_libs 🔒

rmeta 🔒

Structs

EncodedMetadata

Constants

METADATA_HEADER

Metadata header which includes METADATA_VERSION.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

encode_metadata

find_native_static_library

provide

provide_extern