Crate rustc_metadata

Re-exports

pub use creader::DylibError;

pub use creader::load_symbol_from_dylib;

pub use fs::METADATA_FILENAME;

pub use fs::emit_wrapper_file;

Modules

creader

Validates all used crates and extern libraries and loads their metadata

dependency_format 🔒

Resolution of mixing rlibs and dylibs

errors

foreign_modules 🔒

fs

locator

Finds crate binaries and loads their metadata

native_libs 🔒

rmeta 🔒

Structs

EncodedMetadata

Constants

METADATA_HEADER

Metadata header which includes METADATA_VERSION.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

encode_metadata

find_native_static_library

provide

rendered_const

Build a textual representation of an unevaluated constant expression.

try_find_native_dynamic_library

try_find_native_static_library

walk_native_lib_search_dirs