Struct rustc_metadata::creader::CStore

pub struct CStore {
    metadata_loader: Box<MetadataLoaderDyn>,
    metas: IndexVec<CrateNum, Option<Box<CrateMetadata>>>,
    injected_panic_runtime: Option<CrateNum>,
    allocator_kind: Option<AllocatorKind>,
    alloc_error_handler_kind: Option<AllocatorKind>,
    has_global_allocator: bool,
    has_alloc_error_handler: bool,
    pub(crate) stable_crate_ids: StableCrateIdMap,
    unused_externs: Vec<Symbol>,
}

Fields
metadata_loader: Box<MetadataLoaderDyn>metas: IndexVec<CrateNum, Option<Box<CrateMetadata>>>injected_panic_runtime: Option<CrateNum>allocator_kind: Option<AllocatorKind>

This crate needs an allocator and either provides it itself, or finds it in a dependency. If the above is true, then this field denotes the kind of the found allocator.

alloc_error_handler_kind: Option<AllocatorKind>

This crate needs an allocation error handler and either provides it itself, or finds it in a dependency. If the above is true, then this field denotes the kind of the found allocator.

has_global_allocator: bool

This crate has a #[global_allocator] item.

has_alloc_error_handler: bool

This crate has a #[alloc_error_handler] item.

stable_crate_ids: StableCrateIdMap

The interned StableCrateIds.

unused_externs: Vec<Symbol>

Unused externs of the crate

Implementations

impl CStore

pub fn ctor_untracked(&self, def: DefId) -> Option<(CtorKind, DefId)>

pub fn load_macro_untracked(&self, id: DefId, tcx: TyCtxt<'_>) -> LoadedMacro

pub fn def_span_untracked(&self, def_id: DefId, sess: &Session) -> Span

pub fn def_kind_untracked(&self, def: DefId) -> DefKind

pub fn expn_that_defined_untracked( &self, def_id: DefId, sess: &Session ) -> ExpnId

pub fn num_def_ids_untracked(&self, cnum: CrateNum) -> usize

Only public-facing way to traverse all the definitions in a non-local crate. Critically useful for this third-party project: https://github.com/hacspec/hacspec. See https://github.com/rust-lang/rust/pull/85889 for context.

pub fn get_proc_macro_quoted_span_untracked( &self, cnum: CrateNum, id: usize, sess: &Session ) -> Span

pub fn set_used_recursively(&mut self, cnum: CrateNum)

pub(crate) fn update_extern_crate( &mut self, cnum: CrateNum, extern_crate: ExternCrate )

impl CStore

pub fn from_tcx(tcx: TyCtxt<'_>) -> FreezeReadGuard<'_, CStore>

pub fn from_tcx_mut(tcx: TyCtxt<'_>) -> FreezeWriteGuard<'_, CStore>

fn intern_stable_crate_id( &mut self, root: &CrateRoot ) -> Result<CrateNum, CrateError>

pub fn has_crate_data(&self, cnum: CrateNum) -> bool

pub(crate) fn get_crate_data(&self, cnum: CrateNum) -> CrateMetadataRef<'_>

pub(crate) fn get_crate_data_mut( &mut self, cnum: CrateNum ) -> &mut CrateMetadata

fn set_crate_data(&mut self, cnum: CrateNum, data: CrateMetadata)

pub(crate) fn iter_crate_data( &self ) -> impl Iterator<Item = (CrateNum, &CrateMetadata)>

fn iter_crate_data_mut( &mut self ) -> impl Iterator<Item = (CrateNum, &mut CrateMetadata)>

fn push_dependencies_in_postorder( &self, deps: &mut Vec<CrateNum>, cnum: CrateNum )

pub(crate) fn crate_dependencies_in_postorder( &self, cnum: CrateNum ) -> Vec<CrateNum>

fn crate_dependencies_in_reverse_postorder( &self, cnum: CrateNum ) -> Vec<CrateNum>

pub(crate) fn injected_panic_runtime(&self) -> Option<CrateNum>

pub(crate) fn allocator_kind(&self) -> Option<AllocatorKind>

pub(crate) fn alloc_error_handler_kind(&self) -> Option<AllocatorKind>

pub(crate) fn has_global_allocator(&self) -> bool

pub(crate) fn has_alloc_error_handler(&self) -> bool

pub fn report_unused_deps(&self, tcx: TyCtxt<'_>)

pub fn new( metadata_loader: Box<MetadataLoaderDyn>, local_stable_crate_id: StableCrateId ) -> CStore