Struct rustc_metadata::creader::CrateMetadataRef

pub(crate) struct CrateMetadataRef<'a> {
    pub cdata: &'a CrateMetadata,
    pub cstore: &'a CStore,
}

A reference to CrateMetadata that can also give access to whole crate store when necessary.

Fields
cdata: &'a CrateMetadatacstore: &'a CStore
Implementations

impl<'a, 'tcx> CrateMetadataRef<'a>

fn missing(self, descr: &str, id: DefIndex) -> !

fn raw_proc_macro(self, id: DefIndex) -> &'a ProcMacro

fn opt_item_name(self, item_index: DefIndex) -> Option<Symbol>

fn item_name(self, item_index: DefIndex) -> Symbol

fn opt_item_ident(self, item_index: DefIndex, sess: &Session) -> Option<Ident>

fn item_ident(self, item_index: DefIndex, sess: &Session) -> Ident

pub(super) fn map_encoded_cnum_to_current(self, cnum: CrateNum) -> CrateNum

fn def_kind(self, item_id: DefIndex) -> DefKind

fn get_span(self, index: DefIndex, sess: &Session) -> Span

fn load_proc_macro(self, id: DefIndex, tcx: TyCtxt<'tcx>) -> SyntaxExtension

fn get_explicit_item_bounds( self, index: DefIndex, tcx: TyCtxt<'tcx> ) -> EarlyBinder<&'tcx [(Clause<'tcx>, Span)]>

fn get_variant( self, kind: DefKind, index: DefIndex, parent_did: DefId ) -> (VariantIdx, VariantDef)

fn get_adt_def(self, item_id: DefIndex, tcx: TyCtxt<'tcx>) -> AdtDef<'tcx>

fn get_visibility(self, id: DefIndex) -> Visibility<DefId>

fn get_trait_item_def_id(self, id: DefIndex) -> Option<DefId>

fn get_expn_that_defined(self, id: DefIndex, sess: &Session) -> ExpnId

fn get_debugger_visualizers(self) -> Vec<DebuggerVisualizerFile>

fn get_lib_features(self, tcx: TyCtxt<'tcx>) -> &'tcx [(Symbol, Option<Symbol>)]

Iterates over all the stability attributes in the given crate.

fn get_stability_implications( self, tcx: TyCtxt<'tcx> ) -> &'tcx [(Symbol, Symbol)]

Iterates over the stability implications in the given crate (when a #[unstable] attribute has an implied_by meta item, then the mapping from the implied feature to the actual feature is a stability implication).

fn get_lang_items(self, tcx: TyCtxt<'tcx>) -> &'tcx [(DefId, LangItem)]

Iterates over the language items in the given crate.

fn get_stripped_cfg_items( self, cnum: CrateNum, tcx: TyCtxt<'tcx> ) -> &'tcx [StrippedCfgItem]

fn get_diagnostic_items(self) -> DiagnosticItems

Iterates over the diagnostic items in the given crate.

fn get_mod_child(self, id: DefIndex, sess: &Session) -> ModChild

fn get_module_children( self, id: DefIndex, sess: &'a Session ) -> impl Iterator<Item = ModChild> + 'a

Iterates over all named children of the given module, including both proper items and reexports. Module here is understood in name resolution sense - it can be a mod item, or a crate root, or an enum, or a trait.

fn is_ctfe_mir_available(self, id: DefIndex) -> bool

fn is_item_mir_available(self, id: DefIndex) -> bool

fn get_fn_has_self_parameter(self, id: DefIndex, sess: &'a Session) -> bool

fn get_associated_item_or_field_def_ids( self, id: DefIndex ) -> impl Iterator<Item = DefId> + 'a

fn get_associated_item(self, id: DefIndex, sess: &'a Session) -> AssocItem

fn get_ctor(self, node_id: DefIndex) -> Option<(CtorKind, DefId)>

fn get_item_attrs( self, id: DefIndex, sess: &'a Session ) -> impl Iterator<Item = Attribute> + 'a

fn get_inherent_implementations_for_type( self, tcx: TyCtxt<'tcx>, id: DefIndex ) -> &'tcx [DefId]

fn get_traits(self) -> impl Iterator<Item = DefId> + 'a

Decodes all traits in the crate (for rustdoc and rustc diagnostics).

fn get_trait_impls(self) -> impl Iterator<Item = DefId> + 'a

Decodes all trait impls in the crate (for rustdoc).

fn get_incoherent_impls( self, tcx: TyCtxt<'tcx>, simp: SimplifiedType ) -> &'tcx [DefId]

fn get_implementations_of_trait( self, tcx: TyCtxt<'tcx>, trait_def_id: DefId ) -> &'tcx [(DefId, Option<SimplifiedType>)]

fn get_native_libraries( self, sess: &'a Session ) -> impl Iterator<Item = NativeLib> + 'a

fn get_proc_macro_quoted_span(self, index: usize, sess: &Session) -> Span

fn get_foreign_modules( self, sess: &'a Session ) -> impl Iterator<Item = ForeignModule> + '_

fn get_dylib_dependency_formats( self, tcx: TyCtxt<'tcx> ) -> &'tcx [(CrateNum, LinkagePreference)]

fn get_missing_lang_items(self, tcx: TyCtxt<'tcx>) -> &'tcx [LangItem]

fn exported_symbols( self, tcx: TyCtxt<'tcx> ) -> &'tcx [(ExportedSymbol<'tcx>, SymbolExportInfo)]

fn get_macro(self, id: DefIndex, sess: &Session) -> MacroDef

fn def_key(self, index: DefIndex) -> DefKey

fn def_path(self, id: DefIndex) -> DefPath

fn def_path_hash_unlocked( self, index: DefIndex, def_path_hashes: &mut FxHashMap<DefIndex, DefPathHash> ) -> DefPathHash

fn def_path_hash(self, index: DefIndex) -> DefPathHash

fn def_path_hash_to_def_index(self, hash: DefPathHash) -> DefIndex

fn expn_hash_to_expn_id( self, sess: &Session, index_guess: u32, hash: ExpnHash ) -> ExpnId

fn imported_source_file( self, source_file_index: u32, sess: &Session ) -> ImportedSourceFile

Imports the source_map from an external crate into the source_map of the crate currently being compiled (the “local crate”).

The import algorithm works analogous to how AST items are inlined from an external crate’s metadata: For every SourceFile in the external source_map an ‘inline’ copy is created in the local source_map. The correspondence relation between external and local SourceFiles is recorded in the ImportedSourceFile objects returned from this function. When an item from an external crate is later inlined into this crate, this correspondence information is used to translate the span information of the inlined item so that it refers the correct positions in the local source_map (see <decoder::DecodeContext as SpecializedDecoder<Span>>).

The import algorithm in the function below will reuse SourceFiles already existing in the local source_map. For example, even if the SourceFile of some source file of libstd gets imported many times, there will only ever be one SourceFile object for the corresponding file in the local source_map.

Note that imported SourceFiles do not actually contain the source code of the file they represent, just information about length, line breaks, and multibyte characters. This information is enough to generate valid debuginfo for items inlined from other crates.

Proc macro crates don’t currently export spans, so this function does not have to work for them.

fn get_generator_diagnostic_data( self, tcx: TyCtxt<'tcx>, id: DefIndex ) -> Option<GeneratorDiagnosticData<'tcx>>

fn get_attr_flags(self, index: DefIndex) -> AttrFlags

fn get_is_intrinsic(self, index: DefIndex) -> bool

fn get_doc_link_resolutions(self, index: DefIndex) -> DocLinkResMap

fn get_doc_link_traits_in_scope( self, index: DefIndex ) -> impl Iterator<Item = DefId> + 'a