Struct rustc_metadata::cstore::CrateMetadata

pub struct CrateMetadata {
    pub name: String,
    pub extern_crate: Cell<Option<ExternCrate>>,
    pub blob: MetadataBlob,
    pub cnum_map: RefCell<CrateNumMap>,
    pub cnum: CrateNum,
    pub codemap_import_info: RefCell<Vec<ImportedFileMap>>,
    pub root: CrateRoot,
    pub key_map: FnvHashMap<DefKey, DefIndex>,
    pub explicitly_linked: Cell<bool>,
}

Unstable (rustc_private)

Fields
name: String
Unstable (rustc_private)
extern_crate: Cell<Option<ExternCrate>>
Unstable (rustc_private)

Information about the extern crate that caused this crate to be loaded. If this is None, then the crate was injected (e.g., by the allocator)

blob: MetadataBlob
Unstable (rustc_private)
cnum_map: RefCell<CrateNumMap>
Unstable (rustc_private)
cnum: CrateNum
Unstable (rustc_private)
codemap_import_info: RefCell<Vec<ImportedFileMap>>
Unstable (rustc_private)
root: CrateRoot
Unstable (rustc_private)
key_map: FnvHashMap<DefKey, DefIndex>
Unstable (rustc_private)

For each public item in this crate, we encode a key. When the crate is loaded, we read all the keys and put them in this hashmap, which gives the reverse mapping. This allows us to quickly retrace a DefPath, which is needed for incremental compilation support.

explicitly_linked: Cell<bool>
Unstable (rustc_private)

Flag if this crate is required by an rlib version of this crate, or in other words whether it was explicitly linked to. An example of a crate where this is false is when an allocator crate is injected into the dependency list, and therefore isn't actually needed to link an rlib.

Methods

impl<'a, 'tcx> CrateMetadata

fn get_def(&self, index: DefIndex) -> Option<Def>

Unstable (rustc_private)

fn get_trait_def(&self,
item_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>)
-> TraitDef<'tcx>

Unstable (rustc_private)

fn get_adt_def(&self,
item_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>)
-> AdtDefMaster<'tcx>

Unstable (rustc_private)

fn get_predicates(&self,
item_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>)
-> GenericPredicates<'tcx>

Unstable (rustc_private)

fn get_super_predicates(&self,
item_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>)
-> GenericPredicates<'tcx>

Unstable (rustc_private)

fn get_generics(&self,
item_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>)
-> Generics<'tcx>

Unstable (rustc_private)

fn get_type(&self, id: DefIndex, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> Ty<'tcx>

Unstable (rustc_private)

fn get_stability(&self, id: DefIndex) -> Option<Stability>

Unstable (rustc_private)

fn get_deprecation(&self, id: DefIndex) -> Option<Deprecation>

Unstable (rustc_private)

fn get_visibility(&self, id: DefIndex) -> Visibility

Unstable (rustc_private)

fn get_parent_impl(&self, id: DefIndex) -> Option<DefId>

Unstable (rustc_private)

fn get_impl_polarity(&self, id: DefIndex) -> ImplPolarity

Unstable (rustc_private)

fn get_custom_coerce_unsized_kind(&self,
id: DefIndex)
-> Option<CustomCoerceUnsized>

Unstable (rustc_private)

fn get_impl_trait(&self,
id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>)
-> Option<TraitRef<'tcx>>

Unstable (rustc_private)

fn get_lang_items(&self) -> Vec<(DefIndex, usize)>

Unstable (rustc_private)

Iterates over the language items in the given crate.

fn each_child_of_item<F>(&self, id: DefIndex, callback: F) where F: FnMut(Export)

Unstable (rustc_private)

Iterates over each child of the given item.

fn maybe_get_item_ast(&self,
tcx: TyCtxt<'a, 'tcx, 'tcx>,
id: DefIndex)
-> Option<&'tcx InlinedItem>

Unstable (rustc_private)

fn is_item_mir_available(&self, id: DefIndex) -> bool

Unstable (rustc_private)

fn maybe_get_item_mir(&self,
tcx: TyCtxt<'a, 'tcx, 'tcx>,
id: DefIndex)
-> Option<Mir<'tcx>>

Unstable (rustc_private)

fn get_impl_or_trait_item(&self,
id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>)
-> Option<ImplOrTraitItem<'tcx>>

Unstable (rustc_private)

fn get_item_variances(&self, id: DefIndex) -> Vec<Variance>

Unstable (rustc_private)

fn get_ctor_kind(&self, node_id: DefIndex) -> CtorKind

Unstable (rustc_private)

fn get_struct_ctor_def_id(&self, node_id: DefIndex) -> Option<DefId>

Unstable (rustc_private)

fn get_item_attrs(&self, node_id: DefIndex) -> Vec<Attribute>

Unstable (rustc_private)

fn get_struct_field_names(&self, id: DefIndex) -> Vec<Name>

Unstable (rustc_private)

fn get_inherent_implementations_for_type(&self, id: DefIndex) -> Vec<DefId>

Unstable (rustc_private)

fn get_implementations_for_trait(&self,
filter: Option<DefId>,
result: &mut Vec<DefId>)

Unstable (rustc_private)

fn get_trait_of_item(&self, id: DefIndex) -> Option<DefId>

Unstable (rustc_private)

fn get_native_libraries(&self) -> Vec<(NativeLibraryKind, String)>

Unstable (rustc_private)

fn get_dylib_dependency_formats(&self) -> Vec<(CrateNum, LinkagePreference)>

Unstable (rustc_private)

fn get_missing_lang_items(&self) -> Vec<LangItem>

Unstable (rustc_private)

fn get_fn_arg_names(&self, id: DefIndex) -> Vec<Name>

Unstable (rustc_private)

fn get_reachable_ids(&self) -> Vec<DefId>

Unstable (rustc_private)

fn is_const_fn(&self, id: DefIndex) -> bool

Unstable (rustc_private)

fn is_extern_item(&self, id: DefIndex, tcx: TyCtxt<'a, 'tcx, 'tcx>) -> bool

Unstable (rustc_private)

fn is_foreign_item(&self, id: DefIndex) -> bool

Unstable (rustc_private)

fn is_defaulted_trait(&self, trait_id: DefIndex) -> bool

Unstable (rustc_private)

fn is_default_impl(&self, impl_id: DefIndex) -> bool

Unstable (rustc_private)

fn closure_kind(&self, closure_id: DefIndex) -> ClosureKind

Unstable (rustc_private)

fn closure_ty(&self,
closure_id: DefIndex,
tcx: TyCtxt<'a, 'tcx, 'tcx>)
-> ClosureTy<'tcx>

Unstable (rustc_private)

fn def_key(&self, id: DefIndex) -> DefKey

Unstable (rustc_private)

fn def_path(&self, id: DefIndex) -> Option<DefPath>

Unstable (rustc_private)

fn imported_filemaps(&'a self,
local_codemap: &CodeMap)
-> Ref<'a, Vec<ImportedFileMap>>

Unstable (rustc_private)

Imports the codemap from an external crate into the codemap of the crate currently being compiled (the "local crate").

The import algorithm works analogous to how AST items are inlined from an external crate's metadata: For every FileMap in the external codemap an 'inline' copy is created in the local codemap. The correspondence relation between external and local FileMaps is recorded in the ImportedFileMap objects returned from this function. When an item from an external crate is later inlined into this crate, this correspondence information is used to translate the span information of the inlined item so that it refers the correct positions in the local codemap (see <decoder::DecodeContext as SpecializedDecoder<Span>>).

The import algorithm in the function below will reuse FileMaps already existing in the local codemap. For example, even if the FileMap of some source file of libstd gets imported many times, there will only ever be one FileMap object for the corresponding file in the local codemap.

Note that imported FileMaps do not actually contain the source code of the file they represent, just information about length, line breaks, and multibyte characters. This information is enough to generate valid debuginfo for items inlined from other crates.

impl CrateMetadata

fn name(&self) -> &str

Unstable (rustc_private)

fn hash(&self) -> Svh

Unstable (rustc_private)

fn disambiguator(&self) -> &str

Unstable (rustc_private)

fn is_staged_api(&self) -> bool

Unstable (rustc_private)

fn is_allocator(&self) -> bool

Unstable (rustc_private)

fn needs_allocator(&self) -> bool

Unstable (rustc_private)

fn is_panic_runtime(&self) -> bool

Unstable (rustc_private)

fn needs_panic_runtime(&self) -> bool

Unstable (rustc_private)

fn is_compiler_builtins(&self) -> bool

Unstable (rustc_private)

fn is_no_builtins(&self) -> bool

Unstable (rustc_private)

fn panic_strategy(&self) -> PanicStrategy

Unstable (rustc_private)