Struct rustc_metadata::rmeta::decoder::CrateMetadata

pub(crate) struct CrateMetadata {

    blob: MetadataBlob,
    root: CrateRoot,
    trait_impls: FxHashMap<(u32, DefIndex), LazyArray<(DefIndex, Option<SimplifiedType>)>>,
    incoherent_impls: FxHashMap<SimplifiedType, LazyArray<DefIndex>>,
    raw_proc_macros: Option<&'static [ProcMacro]>,
    source_map_import_info: Lock<Vec<Option<ImportedSourceFile>>>,
    def_path_hash_map: DefPathHashMapRef<'static>,
    expn_hash_map: OnceCell<UnhashMap<ExpnHash, ExpnIndex>>,
    alloc_decoding_state: AllocDecodingState,
    def_key_cache: Lock<FxHashMap<DefIndex, DefKey>>,
    def_path_hash_cache: Lock<FxHashMap<DefIndex, DefPathHash>>,
    cnum: CrateNum,
    cnum_map: IndexVec<CrateNum, CrateNum>,
    dependencies: Lock<Vec<CrateNum>>,
    dep_kind: Lock<CrateDepKind>,
    source: Lrc<CrateSource>,
    private_dep: bool,
    host_hash: Option<Svh>,
    hygiene_context: HygieneDecodeContext,
    extern_crate: Lock<Option<ExternCrate>>,

}

Fields
blob: MetadataBlob

The primary crate data - binary metadata blob.

root: CrateRoot

NOTE(eddyb) we pass 'static to a 'tcx parameter because this lifetime is only used behind LazyValue, LazyArray, or LazyTable, and therefore acts like a universal (for<'tcx>), that is paired up with whichever TyCtxt is being used to decode those values.

trait_impls: FxHashMap<(u32, DefIndex), LazyArray<(DefIndex, Option<SimplifiedType>)>>

Trait impl data. FIXME: Used only from queries and can use query cache, so pre-decoding can probably be avoided.

incoherent_impls: FxHashMap<SimplifiedType, LazyArray<DefIndex>>

Inherent impls which do not follow the normal coherence rules.

These can be introduced using either #![rustc_coherence_is_core] or #[rustc_allow_incoherent_impl].

raw_proc_macros: Option<&'static [ProcMacro]>

Proc macro descriptions for this crate, if it’s a proc macro crate.

source_map_import_info: Lock<Vec<Option<ImportedSourceFile>>>

Source maps for code from the crate.

def_path_hash_map: DefPathHashMapRef<'static>

For every definition in this crate, maps its DefPathHash to its DefIndex.

expn_hash_map: OnceCell<UnhashMap<ExpnHash, ExpnIndex>>

Likewise for ExpnHash.

alloc_decoding_state: AllocDecodingState

Used for decoding interpret::AllocIds in a cached & thread-safe manner.

def_key_cache: Lock<FxHashMap<DefIndex, DefKey>>

Caches decoded DefKeys.

def_path_hash_cache: Lock<FxHashMap<DefIndex, DefPathHash>>

Caches decoded DefPathHashes.

cnum: CrateNum

ID of this crate, from the current compilation session’s point of view.

cnum_map: IndexVec<CrateNum, CrateNum>

Maps crate IDs as they are were seen from this crate’s compilation sessions into IDs as they are seen from the current compilation session.

dependencies: Lock<Vec<CrateNum>>

Same ID set as cnum_map plus maybe some injected crates like panic runtime.

dep_kind: Lock<CrateDepKind>

How to link (or not link) this crate to the currently compiled crate.

source: Lrc<CrateSource>

Filesystem location of this crate.

private_dep: bool

Whether or not this crate should be consider a private dependency for purposes of the ‘exported_private_dependencies’ lint

host_hash: Option<Svh>

The hash for the host proc macro. Used to support -Z dual-proc-macro.

hygiene_context: HygieneDecodeContext

Additional data used for decoding HygieneData (e.g. SyntaxContext and ExpnId). Note that we store a HygieneDecodeContext for each CrateMetadat. This is because SyntaxContext ids are not globally unique, so we need to track which ids we’ve decoded on a per-crate basis.

extern_crate: Lock<Option<ExternCrate>>

Information about the extern crate item or path that caused this crate to be loaded. If this is None, then the crate was injected (e.g., by the allocator).

Implementations

impl CrateMetadata

pub(crate) fn new(
sess: &Session,
cstore: &CStore,
blob: MetadataBlob,
root: CrateRoot,
raw_proc_macros: Option<&'static [ProcMacro]>,
cnum: CrateNum,
cnum_map: IndexVec<CrateNum, CrateNum>,
dep_kind: CrateDepKind,
source: CrateSource,
private_dep: bool,
host_hash: Option<Svh>
) -> CrateMetadata

pub(crate) fn dependencies(&self) -> LockGuard<'_, Vec<CrateNum>>

pub(crate) fn add_dependency(&self, cnum: CrateNum)

pub(crate) fn update_extern_crate(&self, new_extern_crate: ExternCrate) -> bool

pub(crate) fn source(&self) -> &CrateSource

pub(crate) fn dep_kind(&self) -> CrateDepKind

pub(crate) fn update_dep_kind(
&self,
f: impl FnOnce(CrateDepKind) -> CrateDepKind
)

pub(crate) fn required_panic_strategy(&self) -> Option<PanicStrategy>

pub(crate) fn needs_panic_runtime(&self) -> bool

pub(crate) fn is_panic_runtime(&self) -> bool

pub(crate) fn is_profiler_runtime(&self) -> bool

pub(crate) fn needs_allocator(&self) -> bool

pub(crate) fn has_global_allocator(&self) -> bool

pub(crate) fn has_default_lib_allocator(&self) -> bool

pub(crate) fn is_proc_macro_crate(&self) -> bool

pub(crate) fn name(&self) -> Symbol

pub(crate) fn stable_crate_id(&self) -> StableCrateId

pub(crate) fn hash(&self) -> Svh

fn num_def_ids(&self) -> usize

fn local_def_id(&self, index: DefIndex) -> DefId

fn reverse_translate_def_id(&self, did: DefId) -> Option<DefId>